ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
web3==6.11.3
python-dotenv==1.0.0
redis==5.0.1
httpx==0.25.2

